Changes with v0.2
Made the mainfrm a FixedSingle, so it is no longer possible to resize
it, this is so i could add the webbrowser control to the mainfrm. Added
shipping cost to the free or "Gratis" items, also correctly sorts these
by price along with the normal items. This also makes Form2 which was
just a webbrowser control unnecessary, thus removed.
Release has been made.